pyproject.toml
[tool.poetry]
name = "sheepfish_test_task"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.2.3"
environs = "^9.5.0"
celery = {extras = ["redis"], version = "^5.3.1"}
psycopg2 = "^2.9.6"
djangorestframework = "^3.14.0"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
django-stubs = "^4.2.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
